mqt-ddsim/requirements.txt
loguru
mqt-ddsim~=1.20.2
